Move oci/wclayer to internal for now
We may want to iterate on this interface a bit before we support it
outside of hcsshim.
